% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMP-methods.R
\name{as.data.frame.PMP}
\alias{as.data.frame.PMP}
\title{Convert a PMP object into a data frame}
\usage{
\method{as.data.frame}{PMP}(x, ...)
}
\arguments{
\item{x}{valid \code{\link{PMP}} object}

\item{...}{arguments to be passed to \code{data.frame}}
}
\value{
a \code{data.frame} with the posterior and prior probabilities as
well as the definition of the models
}
\description{
This function takes a PMP object an returns a \code{data.frame} summarising
the information.
}
\author{
Rachel Heyard
}
