% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVS_plot.R
\name{plot_CSVS}
\alias{plot_CSVS}
\title{Plot a CSVS object}
\usage{
plot_CSVS(CSVSobject, namesVar = NULL, shrunken = FALSE,
  standardized = FALSE, numberIntercepts, ...)
}
\arguments{
\item{CSVSobject}{valid \code{\link{CSVS}} object}

\item{namesVar}{names of the variables}

\item{shrunken}{should the coefficients be shrunken?}

\item{standardized}{should the coefficients be standardized?}

\item{numberIntercepts}{how many cause-specific intercepts are in the
model for each outcome}

\item{...}{parameters for plot}
}
\description{
Plot a CSVS object
}
\author{
Rachel Heyard
}
