% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_multinomial.R
\name{sample_multinomial}
\alias{sample_multinomial}
\title{Samples from a PMP object}
\usage{
sample_multinomial(PMP_object, shrink = TRUE, data, which = "MPM",
  discreteSurv = TRUE)
}
\arguments{
\item{PMP_object}{formula of the model including all potential variables}

\item{shrink}{should the coefficients be shrunken towards their prior mean?}

\item{data}{the (training) data frame with all the information}

\item{which}{which model should be sampled from? either an integer, 'MPM'
or 'MAP'}

\item{discreteSurv}{Boolean variable telling us whether a 'simple'
multinomial regression is looked for or if the goal is a discrete
survival-time model for multiple modes of failure is needed.}
}
\value{
returns an object with the model coefficients and supplementary
information
}
\description{
This function samples from a specific model inside a \code{PMP} object.
}
\author{
Rachel Heyard
}
