% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBFmultinomial-package.R
\docType{package}
\name{TBFmultinomial-package}
\alias{TBFmultinomial-package}
\alias{TBFmultinomial}
\title{Objective Bayesian variable selection for multinomial regression and
discrete time-to-event models with competing risks}
\description{
Extension of the TBF methdology introduced by Held et al. (2015) <doi:10.1214/14-STS510> to discrete time-to-event models with competing risks (also applicable to the multinomial regression model)
}
\author{
Rachel Heyard \email{rachel.heyard@uzh.ch}
}
\keyword{package}
