% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAP_data.R
\docType{data}
\name{VAP_data}
\alias{VAP_data}
\title{Data on VAP acquistion in one ICU}
\format{A data frame with 1640 rows and 7 variables on 90 distinct patients:
 \describe{
  \item{ID}{distinct ID for each patient}
  \item{day}{day of ventilation, day = 1 is the first day of ventilation}
  \item{type}{is it a medical or a surgical patient}
  \item{gender}{gender of the patient, 1 = male, 0 = female}
  \item{SAPSadmission}{the SAPS 2 score at admission to the ICU}
  \item{SOFA}{the daily SOFA score}
  \item{outcome}{final outcome after the first observation period}
}}
\usage{
data(VAP_data)
}
\description{
It is a tiny subset of the OUTCOMEREA database whose only perhaps will be to
test an illustrate the functions of this package.
}
